+++ gorm-tools/src/main/groovy/gorm/tools/problem/ProblemHandler.groovy
@@ -41,6 +41,10 @@ class ProblemHandler {
 
     @Autowired ICUMessageSource messageSource
 
+    static {
+        stackTraceUtilsDefaultFilters()
+    }
+
     GenericProblem handleException(Class entityClass, Throwable e) {
         handleException(e, entityClass.simpleName)
     }
@@ -107,7 +111,13 @@ class ProblemHandler {
         }
         else if (e instanceof ThrowableProblem) {
             return (GenericProblem) e.problem
-        } else {
+        }
+        else if (e instanceof NullPointerException) {
+            //deal with the dreaded null pointer
+            String stackLine1 = e.stackTrace[0].toString()
+            return new UnexpectedProblem().cause(e).detail("NullPointerException at ${stackLine1}")
+        }
+        else {
             return new UnexpectedProblem().cause(e).detail(e.message)
         }
     }
@@ -175,8 +185,8 @@ class ProblemHandler {
      * Broken pipe exception happens when client has closed the socket and server tries to write/send any response byte on the output stream.
      * Server Can write nothing to output stream once we encounter Broken pipe exception
      */
-    static isBrokenPipe(Exception ex) {
-        return ex.message.toLowerCase().contains("broken pipe")
+    static boolean isBrokenPipe(Exception ex) {
+        return ex.message && ex.message.toLowerCase().contains("broken pipe")
     }
 
     //Legacy from ValidationException
@@ -195,5 +205,48 @@ class ProblemHandler {
         }
         return b.toString();
     }
+
+    @SuppressWarnings(['BooleanMethodReturnsNull'])
+    static void stackTraceUtilsDefaultFilters(){
+        StackTraceUtils.addClassTest { String className ->
+            for (String groovyPackage : (NOISY_PACKAGES + NOISY_TEST_PACKAGES)) {
+                if (className.startsWith(groovyPackage)) {
+                    return false
+                }
+            }
+            return null
+        }
+    }
+
+    //the list of packages to summarize up so logging trace is not so noisy. only logs one line if multiples start with these
+    public static List NOISY_PACKAGES = [
+        'jdk.internal.reflect.NativeMethodAccessorImpl',
+        'jdk.internal.reflect.DelegatingMethodAccessorImpl',
+        'jdk.internal.reflect.GeneratedMethodAccessor',
+        'org.springframework.web.filter.OncePerRequestFilter',
+        'org.springframework.web.filter.CharacterEncodingFilter',
+        'org.springframework.web.filter.DelegatingFilterProxy',
+        'org.springframework.security.web',
+        'org.grails.core.DefaultGrailsControllerClass',
+        'org.grails.web.servlet.mvc.GrailsWebRequestFilter',
+        'org.grails.web.filters.HiddenHttpMethodFilter',
+        'org.grails.datastore.mapping.reflect.FieldEntityAccess',
+        'org.apache.catalina.core',
+        'org.apache.tomcat.websocket.server.WsFilter',
+        'org.apache.tomcat.util.net',
+        'org.apache.tomcat.util.threads',
+        'org.apache.coyote'
+    ];
+
+    public static List NOISY_TEST_PACKAGES = [
+        'jdk.internal.reflect.NativeConstructorAccessorImpl',
+        'org.spockframework.runtime',
+        'org.spockframework.util.ReflectionUtil',
+        'org.spockframework.junit4.ExceptionAdapterInterceptor',
+        'org.junit.platform.engine.support.hierarchical',
+        'org.junit.platform.launcher.core.EngineExecutionOrchestrator',
+        //'org.gradle',
+
+    ];
 
 }