+++ gorm-tools/src/main/groovy/gorm/tools/repository/GormRepo.groovy
@@ -32,6 +32,7 @@ import gorm.tools.transaction.TrxUtils
 import gorm.tools.utils.GormMetaUtils
 import grails.core.support.proxy.ProxyHandler
 import grails.validation.ValidationException
+import yakworks.api.problem.ThrowableProblem
 import yakworks.api.problem.data.NotFoundProblem
 import yakworks.commons.lang.ClassUtils
 
@@ -259,19 +260,22 @@ trait GormRepo<D> implements BulkableRepo<D>, QueryMangoEntityApi<D> {
         if(ident){
             //return it fast if its good to go, will have blown and error if not found
             return get(ident, data['version'] as Long)
-        } else if(Lookupable.isAssignableFrom(getEntityClass())){
+        }
+        if(Lookupable.isAssignableFrom(getEntityClass())){
             // call the lookup if domain implements the Lookupable
             //FIXME is there a cleaner way to do this?
             foundEntity = (D) ClassUtils.callStaticMethod(getEntityClass(), 'lookup', data)
-        } else {
+        }
+        if(!foundEntity) {
             foundEntity = lookup(data)
         }
-        if(mustExist) RepoUtil.checkFound(foundEntity, 'data map', getEntityClass().name)
+        if(mustExist) RepoUtil.checkFound(foundEntity, data as Serializable, getEntityClass().name)
         return foundEntity
     }
 
     /**
      * does nothing, can be implemented by the conrete repo for special lookup logic such as for souceId
+     * if nothing is found it should return null or throw a DataRetrievalFailureException if message needed
      */
     D lookup(Map data) {
         return null
@@ -347,7 +351,7 @@ trait GormRepo<D> implements BulkableRepo<D>, QueryMangoEntityApi<D> {
      * Transactional, Calls delete always with flush = true so we can intercept any DataIntegrityViolationExceptions.
      *
      * @param entity the domain entity
-     * @throws DataProblemException if a DataIntegrityViolationException is thrown
+     * @throws ThrowableProblem if a DataIntegrityViolationException is thrown
      */
     void remove(D entity, Map args = [:]) {
         try {