+++ gorm-tools/src/main/groovy/yakworks/gorm/api/ApiUtils.groovy
@@ -37,8 +37,8 @@ class ApiUtils {
 
     /**
      * Parses name=xyz&size=123 query string into a map
-     * @param queryString Query string
-     * @return Map<String,String>
+     * @param queryString the query params string to parse
+     * @return Map<String,String> the map version of the parsed string
      */
     static Map parseQueryParams(String queryString) {
         if(!queryString) return [:]