+++ gorm-tools/src/test/groovy/gorm/tools/repository/BulkableRepoSpec.groovy
@@ -31,8 +31,7 @@ class BulkableRepoSpec extends Specification implements GormHibernateTest {
     @Autowired KitchenSinkRepo kitchenSinkRepo
 
     SyncJobArgs setupSyncJobArgs(DataOp op = DataOp.add){
-        return new SyncJobArgs(parallel: false, op: op, source: "test", sourceId: "test",
-            includes: ["id", "name", "ext.name"])
+        return new SyncJobArgs(parallel: false, async:false, op: op, source: "test", sourceId: "test", includes: ["id", "name", "ext.name"])
     }
 
     void "sanity check bulkable repo"() {