+++ security/boot-security-gorm/src/main/groovy/yakworks/security/gorm/AppUserService.groovy
@@ -12,7 +12,6 @@ import groovy.transform.CompileStatic
 
 import org.grails.datastore.mapping.query.api.Criteria
 import org.springframework.beans.factory.annotation.Autowired
-import org.springframework.beans.factory.annotation.Value
 
 import grails.gorm.transactions.Transactional
 import yakworks.security.gorm.model.AppUser
@@ -29,18 +28,6 @@ import yakworks.security.user.CurrentUser
 @CompileStatic
 @Transactional
 class AppUserService {
-
-    @Value('${yakworks.security.password.historyEnabled:false}')
-    boolean passwordHistoryEnabled
-
-    @Value('${yakworks.security.password.expireDays:90}')
-    int passwordExpireDays
-
-    @Value('${yakworks.security.password.expireEnabled:false}')
-    boolean passwordExpiryEnabled
-
-    @Value('${yakworks.security.password.warnDays:30}')
-    int passwordWarnDays
 
     @Autowired
     CurrentUser currentUser
@@ -69,13 +56,13 @@ class AppUserService {
         } as List
         if (secLoginHistoryList) {
             SecLoginHistory secLoginHistory = secLoginHistoryList[0]
-            secLoginHistory.logoutDate = new Date()
+            secLoginHistory.logoutDate = LocalDateTime.now()
             secLoginHistory.save()
         }
     }
 
     Integer remainingDaysForPasswordExpiry(AppUser u) {
-        LocalDateTime pExpire = u.passwordChangedDate.plusDays(passwordExpireDays)
+        LocalDateTime pExpire = u.passwordChangedDate.plusDays(passwordValidator.passwordExpireDays)
         return Duration.between(LocalDateTime.now(), pExpire).toDays().toInteger()
     }
 
@@ -88,7 +75,7 @@ class AppUserService {
         user.passwordChangedDate = LocalDateTime.now()
         user.save()
 
-        if (passwordHistoryEnabled) {
+        if (passwordValidator.passwordHistoryEnabled) {
             SecPasswordHistory.create(user, newPwd)
         }
     }